requirements.txt
torch~=2.0.0
numpy~=1.24.3
scipy~=1.11.1
torchvision~=0.15.1
tqdm~=4.65.0
pandas~=2.0.3
opencv-python~=4.8.0.74
MedPy~=0.4.0
Pillow~=9.5.0
pycocotools~=2.0.6
matplotlib~=3.7.2